+++ src/components/Card/Card.styles.css
@@ -1,9 +1,14 @@
 .card {
   display: flex;
   flex-direction: column;
-  background-color: #dfe6e6;
-  border: 1px solid grey;
-  border-radius: 5px;
+  background-color: white;
   margin: 25px;
   cursor: pointer;
+  box-shadow: 0 0 15px 3px rgba(0, 0, 0, 0.26);
+  border-radius: 18px;
+}
+
+
+.card a{
+  color:  black
 }
