+++ src/components/Navbar/Navbar.styles.css
@@ -6,6 +6,7 @@
   font-family: Helvetica;
   font-weight: 300;
   box-shadow: 0 0 15px 3px rgba(0, 0, 0, 0.56);
+  padding: 5px;
 }
 
 .navbar_title {
