+++ src/pages/Home/Home.page.jsx
@@ -1,20 +1,12 @@
 import React, { useRef } from 'react';
-import { Link, useHistory } from 'react-router-dom';
 import CardList from '../../components/Cardlist/CardList.component';
 
 import { useAuth } from '../../providers/Auth';
 import './Home.styles.css';
 
 function HomePage() {
-  const history = useHistory();
   const sectionRef = useRef(null);
-  const { authenticated, logout } = useAuth();
-
-  function deAuthenticate(event) {
-    event.preventDefault();
-    logout();
-    history.push('/');
-  }
+  const { authenticated } = useAuth();
 
   return (
     <section className="homepage" ref={sectionRef}>
@@ -22,16 +14,9 @@ function HomePage() {
       {authenticated ? (
         <>
           <h2>Good to have you back</h2>
-          <span>
-            <Link to="/" onClick={deAuthenticate}>
-              ← logout
-            </Link>
-            <span className="separator" />
-            <Link to="/secret">show me something cool →</Link>
-          </span>
         </>
       ) : (
-        <Link to="/login">let me in →</Link>
+        <> </>
       )}
       <CardList />
     </section>
