+++ src/pages/Player/Player.styles.css
@@ -1,0 +1,12 @@
+.player_page_container {
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  margin-top: -579px;
+  padding-left: 3px;
+}
+
+.player_container {
+  width: 70%;
+  background-color: black;
+}
